refactor : move global shared models to core.shared package
