+++ src/main/java/coz/weavon/core/auth/presentation/controller/AuthController.java
@@ -1,7 +1,7 @@
 package coz.weavon.core.auth.presentation.controller;
 
 import coz.weavon.helper.MessageTranslator;
-import coz.weavon.common.presentation.model.response.RestResponse;
+import coz.weavon.core.shared.presentation.model.response.RestResponse;
 import coz.weavon.core.auth.application.service.AuthService;
 import coz.weavon.core.auth.domain.model.AuthUser;
 import coz.weavon.core.auth.presentation.model.AuthJoinRequest;
