+++ src/main/java/coz/weavon/core/auth/presentation/filter/UsernameAuthenticationFilter.java
@@ -2,8 +2,8 @@ package coz.weavon.core.auth.presentation.filter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import coz.weavon.helper.MessageTranslator;
-import coz.weavon.common.presentation.model.response.ErrorResponse;
-import coz.weavon.common.presentation.model.response.RestResponse;
+import coz.weavon.core.shared.presentation.model.response.ErrorResponse;
+import coz.weavon.core.shared.presentation.model.response.RestResponse;
 import coz.weavon.core.auth.domain.model.AuthToken;
 import coz.weavon.core.auth.domain.model.AuthUser;
 import jakarta.servlet.FilterChain;
