+++ src/main/java/coz/weavon/core/auth/presentation/handler/OAuthFailureHandler.java
@@ -2,8 +2,8 @@ package coz.weavon.core.auth.presentation.handler;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import coz.weavon.helper.MessageTranslator;
-import coz.weavon.common.presentation.model.response.ErrorResponse;
-import coz.weavon.common.presentation.model.response.RestResponse;
+import coz.weavon.core.shared.presentation.model.response.ErrorResponse;
+import coz.weavon.core.shared.presentation.model.response.RestResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
