+++ src/main/java/coz/weavon/core/shared/application/model/condition/RestCondition.java
@@ -1,4 +1,4 @@
-package coz.weavon.common.application.model.condition;
+package coz.weavon.core.shared.application.model.condition;
 
 import java.util.Collection;
 import java.util.Objects;
