+++ src/main/java/coz/weavon/core/shared/presentation/model/response/RestResponse.java
@@ -1,4 +1,4 @@
-package coz.weavon.common.presentation.model.response;
+package coz.weavon.core.shared.presentation.model.response;
 
 import lombok.Builder;
 import lombok.Getter;
