+++ src/main/java/coz/weavon/core/user/application/model/condition/UserSearchCondition.java
@@ -1,7 +1,7 @@
 package coz.weavon.core.user.application.model.condition;
 
 import com.querydsl.core.types.dsl.BooleanExpression;
-import coz.weavon.common.application.model.condition.RestCondition;
+import coz.weavon.core.shared.application.model.condition.RestCondition;
 import coz.weavon.exception.model.BusinessException;
 import coz.weavon.core.user.application.model.command.UserSearchCommand;
 import coz.weavon.core.user.infrastructure.model.QUserEntity;
